[python] extract converter_binop.cpp and converter_unop.cpp
Move all 13 binary-operator methods plus the unary operator method of
python_converter into two new TUs:

  converter/converter_binop.cpp: get_logical_operator_expr,
  convert_function_calls_to_side_effects, handle_chained_comparisons_logic,
  handle_membership_operator, get_binary_operator_expr,
  handle_none_check_setup, handle_array_operations, handle_tuple_operations,
  handle_list_operations, handle_relational_type_mismatches,
  handle_string_binary_operations, build_binary_expression,
  promote_ieee_operands; plus three file-local helpers (is_ieee_op,
  attach_symbol_location, handle_float_vs_string).

  converter/converter_unop.cpp: get_unary_operator_expr.

Promote map_operator and operator_map into the converter/converter_internal.h
namespace so both new TUs and the remaining python_converter.cpp share the
same inline definitions.

Pure code movement — no logic change.
