+++ src/python-frontend/CMakeLists.txt
@@ -13,11 +13,13 @@
 add_library(pythonfrontend STATIC
             python_language.cpp
             python_converter.cpp
+            converter/converter_binop.cpp
             converter/converter_class.cpp
             converter/converter_compare.cpp
             converter/converter_dunder.cpp
             converter/converter_symbols.cpp
             converter/converter_types.cpp
+            converter/converter_unop.cpp
             converter/converter_util.cpp
             python_class.cpp
             python_class_builder.cpp
